job-discovery/requirements-notify.txt
# Lightweight dependencies for GitHub Actions (no Playwright, no Flask)
# Install is under 30 seconds on ubuntu-latest
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.1.1
python-dateutil==2.9.0
